app/requirements.txt
fastapi[standard]
celery==5.4.0
redis==5.1.1
python-dotenv==1.0.1
pydantic-settings==2.6.0
pytest==8.3.3
pytest-mock==3.14.0